resolving conversations and changed 'gsub("^%s*(.-)$", "%1")' to 'trimSpace()' in talkactions.
