+++ data/scripts/talkactions/god/manage_badge.lua
@@ -21,8 +21,7 @@ function addBadge.onSay(player, words, param)
 		return true
 	end
 
-	-- Trim left
-	split[2] = split[2]:gsub("^%s*(.-)$", "%1")
+	split[2] = split[2]:trimSpace() -- Trim left
 	local id = tonumber(split[2])
 	target:addBadge(id)
 	return true
