+++ data/scripts/talkactions/god/manage_storage.lua
@@ -18,8 +18,7 @@ function Player.getStorageValueTalkaction(self, param)
 		return true
 	end
 
-	-- Trim left
-	split[2] = split[2]:gsub("^%s*(.-)$", "%1")
+	split[2] = split[2]:trimSpace() -- Trim left
 
 	-- Try to convert the second parameter to a number. If it's not a number, treat it as a storage name
 	local storageKey = tonumber(split[2])
