+++ src/game/game_definitions.hpp
@@ -61,30 +61,6 @@ enum LightState_t {
 	LIGHT_STATE_SUNRISE,
 };
 
-enum CyclopediaBadge_t : uint8_t {
-	ACCOUNT_AGE = 1,
-	LOYALTY,
-	ACCOUNT_ALL_LEVEL,
-	ACCOUNT_ALL_VOCATIONS,
-	TOURNAMENT_PARTICIPATION,
-	TOURNAMENT_POINTS,
-};
-
-enum CyclopediaTitle_t : uint8_t {
-	NOTHING = 0,
-	GOLD,
-	MOUNTS,
-	OUTFITS,
-	LEVEL,
-	HIGHSCORES,
-	BESTIARY,
-	LOGIN,
-	TASK,
-	MAP,
-	QUEST,
-	OTHERS,
-};
-
 enum CyclopediaCharacterInfoType_t : uint8_t {
 	CYCLOPEDIA_CHARACTERINFO_BASEINFORMATION = 0,
 	CYCLOPEDIA_CHARACTERINFO_GENERALSTATS = 1,
